# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/kpi/{database}/list": {
            "post": {
                "description": "Measure KPI of each transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KPI"
                ],
                "summary": "List Data KPI",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/structs.PageView"
                        }
                    },
                    {
                        "type": "string",
                        "description": "database path argument",
                        "name": "database",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structs.CommonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "structs.CommonResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "failed": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "resultCount": {
                    "type": "integer"
                },
                "statusCode": {
                    "type": "integer"
                },
                "success": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "structs.PageView": {
            "type": "object",
            "properties": {
                "endDt": {
                    "type": "string",
                    "example": "2024-05-11"
                },
                "mdn": {
                    "type": "string",
                    "example": "081284088408"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "shift": {
                    "type": "string",
                    "example": "sore"
                },
                "startDt": {
                    "type": "string",
                    "example": "2024-05-11"
                },
                "status": {
                    "type": "integer",
                    "example": 20
                },
                "view": {
                    "type": "integer",
                    "example": 10
                }
            }
        }
    }
}